fix: 챔피언 정보 받아오는 함수 수정
 - champ service에 임시 저장
 - champ repository에 임시 저장
